### 2021-10-26 #3854 @whedon
[PRE REVIEW]: DEVSIM: A TCAD Semiconductor Device Simulator
**Submitting author:** @tcaduser (<a href="http://orcid.org/0000-0002-6783-0194">Juan E. Sanchez</a>)
**Repository:** <a href="https://github.com/devsim/devsim.git" target ="_blank">https://github.com/devsim/devsim.git</a>
**Version:** v1.6.0
