### 2021-10-27 06fde40
fix fedora build and add superlu backward compatibility
