+++ scripts/setup_fedora.sh
@@ -5,6 +5,7 @@ CMAKE="/usr/bin/cmake3"
 CMAKE_CXX_FLAGS=""
 CXX="/usr/bin/g++"
 CC="/usr/bin/gcc"
+PY_INCLUDE="/usr/include/python3.9"
 for TYPE in debug release; do
 ARCH=`uname -m`
 #  for ARCH in i386 x86_64; do
@@ -18,6 +19,9 @@ ARCH=`uname -m`
       -DCMAKE_CXX_FLAGS:STRING="${CMAKE_CXX_FLAGS}" \
       -DDEVSIM_EXTENDED_PRECISION=ON \
       -DPYTHON3=ON \
+      -DPYTHON3_INCLUDE=${PY_INCLUDE} \
       ..)
 #  done
 done
+
+
