+++ src/math/SuperLUData.cc
@@ -214,7 +214,9 @@ bool SuperLUData::LUFactorRealMatrix(CompressedMatrix<DoubleType> *cm, const Dou
   panel_size = sp_ienv(1);
   relax = sp_ienv(2);
 
+#if SUPERLU_MAJOR_VERSION == 5
   GlobalLU_t Glu;
+#endif
 
   if (lutype_ == PEnum::LUType_t::FULL)
   {
