### 2021-11-08 #3854 @lucydot
@jgoldfar @dalonsoa thank-you for agreeing to review.  I'll ask our bot Whedon to start a review issue thread (this is the pre-review thread) - the review guidelines will be listed at the top of that new thread.

We're asking for reviews to be complete within a 4-6 week period. As the review is usually an iterative process (rather than something you do in one sitting) it's best for reviews to start earlier in that timeline - providing the review through November will work well 👍
