### 2021-11-18 3b3a6ce
improve build scripts to hard code paths on configuration time
