+++ scripts/build_msys.sh
@@ -3,8 +3,13 @@ set -e
 export PATH=/mingw64/bin:/usr/bin:${PATH}
 # msys Specific
 #pacman -Su --noconfirm rsync zip
+pacman -Su --noconfirm make
 
 # handle miniconda in appveyor.yml
+export CXX=g++
+export PYTHON3_BIN=python
+export PYTHON3_INCLUDE=$(python -c "from sysconfig import get_paths as gp; print(gp()['include'])")
+export PYTHON3_ARCHIVE=$(cygpath -w ${CONDA_PREFIX}/libs/python3.lib)
 
 # this script assumes git clone and submodule initialization has been done
 
