+++ scripts/build_centos_7.sh
@@ -19,10 +19,14 @@ if [ ! -f Miniconda3-latest-Linux-x86_64.sh ]
 then
 curl -L -O https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh;
 bash Miniconda3-latest-Linux-x86_64.sh -b -p ${HOME}/anaconda;
-${HOME}/anaconda/bin/conda create  -y --name python3_devsim_build python=3.8
+${HOME}/anaconda/bin/conda create  -y --name python3_devsim_build python=3
 ${HOME}/anaconda/bin/conda install -y --name python3_devsim_build mkl mkl-devel mkl-include boost cmake
 fi
 source ${HOME}/anaconda/bin/activate python3_devsim_build
+
+export PYTHON3_BIN=python
+export PYTHON3_INCLUDE=$(python -c "from sysconfig import get_paths as gp; print(gp()['include'])")
+export PYTHON3_ARCHIVE=""
 
 #download boost library
 #(cd ${HOME}/devsim/external && curl -O -L https://dl.bintray.com/boostorg/release/1.66.0/source/boost_1_66_0.tar.gz)
