+++ src/main/CMakeLists.txt
@@ -44,6 +44,9 @@ ENDIF(WIN32)
 IF (APPLE)
 SET_TARGET_PROPERTIES(devsim_py3 PROPERTIES LINK_FLAGS "-undefined dynamic_lookup -flat_namespace")
 ENDIF (APPLE)
+IF (${CMAKE_SYSTEM_NAME} MATCHES "Linux")
+SET_TARGET_PROPERTIES(devsim_py3 PROPERTIES INSTALL_RPATH "\$ORIGIN/" BUILD_RPATH "\$ORIGIN/")
+ENDIF (${CMAKE_SYSTEM_NAME} MATCHES "Linux")
 ENDIF (PYTHON3)
 
 IF (PYTHON3)

[…]
