+++ dist/package_macos.sh
@@ -77,6 +77,14 @@ do
 echo install_name_tool -change $j "@loader_path/`basename $j`" $i
 install_name_tool -change $j "@loader_path/`basename $j`" $i
 done
+
+# MKL is now versioned
+for i in ${DIST_PYDLL}/devsim_py*.so
+do
+cp -vf "${CONDA_PREFIX}/lib/libmkl_rt.1.dylib" ${DIST_PYDLL}
+echo install_name_tool -change "@rpath/libmkl_rt.1.dylib" "@loader_path/libmkl_rt.1.dylib" ${DIST_PYDLL}/
+install_name_tool -change "@rpath/libmkl_rt.1.dylib" "@loader_path/libmkl_rt.1.dylib" ${DIST_PYDLL}/
+done
 done
 
 
