+++ dist/package_msys.sh
@@ -87,7 +87,7 @@ cp /mingw64/bin/libwinpthread-1.dll ${DIST_LIB}/symdiff/
 cp -v __init__.py ${DIST_PYDLL}
 
 # Intel MKL
-cp -v ${CONDA_PREFIX}/Library/bin/mkl_rt.1.dll ${DIST_PYDLL}
+#cp -v ${CONDA_PREFIX}/Library/bin/mkl_rt.1.dll ${DIST_PYDLL}
 
 zip -r ${DIST_VER}.zip ${DIST_DIR}
 
