+++ scripts/setup_centos_6.sh
@@ -13,6 +13,8 @@ ARCH=`uname -m`
       -DDEVSIM_EXTENDED_PRECISION=ON \
       -DMKL_PARDISO=ON \
       -DPYTHON3=ON \
+      -DPYTHON3_INCLUDE=${PYTHON3_INCLUDE} \
+      -DCONDA_PREFIX=${CONDA_PREFIX} \
       ..)
 #  done
 done
