### 2021-11-19 e7f6b1a
more work on scripts
