+++ scripts/build_macos.sh
@@ -39,6 +39,10 @@ else
   exit 1;
 fi
 
+export PYTHON3_BIN=python
+export PYTHON3_INCLUDE=$(python -c "from sysconfig import get_paths as gp; print(gp()['include'])")
+export PYTHON3_ARCHIVE=""
+
 #minimal conda environments to prevent linking against the wrong libraries
 if [ "${1}" = "gcc" ] && [ ! -f Miniconda3-latest-MacOSX-x86_64.sh ]
 then
