+++ scripts/setup_appveyor.sh
@@ -6,6 +6,16 @@ AOPTION="$2"
 BUILDDIR="$3"
 CONDA_PREFIX="$4"
 
-mkdir -p ${BUILDDIR}
-(cd ${BUILDDIR}; cmake -G "${1}" -A $2 -DDEVSIM_CONFIG=${DEVSIM_CONFIG} -DPYTHON3=ON  -DMKL_PARDISO=ON -DANACONDA_PATH=${CONDA_PREFIX} -DBUILDDIR=${BUILDDIR} ..)
+    mkdir -p ${BUILDDIR}
+    (cd ${BUILDDIR};
+    cmake -G "${1}" \
+    -A $2 \
+    -DDEVSIM_CONFIG=${DEVSIM_CONFIG} \
+    -DMKL_PARDISO=ON \
+    -DCONDA_PREFIX=${CONDA_PREFIX} \
+    -DPYTHON3=ON \
+    -DPYTHON3_INCLUDE=${PYTHON3_INCLUDE} \
+    -DPYTHON3_ARCHIVE=${PYTHON3_ARCHIVE} \
+    -DBUILDDIR=${BUILDDIR} \
+    ..)
 
