### 2021-11-20 c7d54d9
Reduce verbosity of circuit information during solve.  Update tests to print information manually.
