+++ testing/ssac_cap_2d_element.py
@@ -14,6 +14,7 @@
 
 # cap device to test displacement current at contact
 import devsim
+import test_common
 
 device="MyDevice"
 region="MyRegion"
@@ -142,8 +143,11 @@ print(devsim.get_contact_charge(device=device, contact="top", equation="Potentia
 print(devsim.get_contact_charge(device=device, contact="bot", equation="PotentialEquation"))
 #
 devsim.solve(type="ac", frequency=1e-3)
+test_common.print_ac_circuit_solution()
 devsim.solve(type="ac", frequency=1e10)
+test_common.print_ac_circuit_solution()
 devsim.solve(type="ac", frequency=1e15)
+test_common.print_ac_circuit_solution()
 
 #for i in devsim.get_element_model_values(device=device, region=region, name="DField"):
 #  print(i)
