+++ testing/ssac_cap.py
@@ -94,10 +94,15 @@ devsim.circuit_element(name="R1", n1="topbias", n2=1, value=1e3)
 
 #
 devsim.solve(type="dc", absolute_error=1.0, relative_error=1e-10, maximum_iterations=30)
+
+test_common.print_circuit_solution()
 #
 print(devsim.get_contact_charge(device=device, contact="top", equation="PotentialEquation"))
 print(devsim.get_contact_charge(device=device, contact="bot", equation="PotentialEquation"))
 #
 devsim.solve(type="ac", frequency=1e10)
+test_common.print_ac_circuit_solution()
+
 devsim.solve(type="ac", frequency=1e15)
+test_common.print_ac_circuit_solution()
 
