+++ cmake/ubuntu_18.04.cmake
@@ -21,8 +21,7 @@ SET (SQLITE3_INCLUDE /usr/include)
 
 SET (SYMDIFF_ARCHIVE ${EXTERNAL_LIB}/symdiff/lib/libsymdiff_dynamic.a)
 SET (SYMDIFF_INCLUDE ${EXTERNAL_LIB}/symdiff/include)
-SET (PYTHON3_INCLUDE   /usr/include/python3.6m)
-SET (PYTHON3_ARCHIVE -L /usr/lib/x86_64-linux-gnu -lpython3.6m)
 
 # important flag for dynamic linking of static archives on linux
 SET (CMAKE_POSITION_INDEPENDENT_CODE ON)
+
