+++ scripts/setup_fedora.sh
@@ -1,11 +1,6 @@
 #!/bin/bash
 
 DEVSIM_CONFIG="fedora"
-CMAKE="/usr/bin/cmake3"
-CMAKE_CXX_FLAGS=""
-CXX="/usr/bin/g++"
-CC="/usr/bin/gcc"
-PY_INCLUDE="/usr/include/python3.9"
 for TYPE in debug release; do
 ARCH=`uname -m`
 #  for ARCH in i386 x86_64; do
@@ -19,7 +14,7 @@ ARCH=`uname -m`
       -DCMAKE_CXX_FLAGS:STRING="${CMAKE_CXX_FLAGS}" \
       -DDEVSIM_EXTENDED_PRECISION=ON \
       -DPYTHON3=ON \
-      -DPYTHON3_INCLUDE=${PY_INCLUDE} \
+      -DPYTHON3_INCLUDE=${PYTHON3_INCLUDE} \
       ..)
 #  done
 done
