+++ cmake/fedora.cmake
@@ -10,7 +10,7 @@ SET (QUADMATH_ARCHIVE "-lquadmath")
 # to prevent linking against bad libstdc++
 
 SET (BLAS_ARCHIVE
- ${EXTERNAL_LIB}/getrf/build/libgetrf.a -Wl,--no-as-needed -lpthread -lmkl_rt -lm -ldl
+ ${EXTERNAL_LIB}/getrf/build/libgetrf.a -Wl,--no-as-needed -lpthread -lblas -lm -ldl
 )
 
 SET (SUPERLU_INCLUDE /usr/include/SuperLU)
@@ -22,8 +22,6 @@ SET (ZLIB_ARCHIVE -lz)
 
 SET (SYMDIFF_INCLUDE ${EXTERNAL_LIB}/symdiff/include)
 SET (SYMDIFF_ARCHIVE ${EXTERNAL_LIB}/symdiff/lib/libsymdiff_dynamic.a)
-
-SET (PYTHON3_INCLUDE   /usr/include/python3.9)
 
 # important flag for dynamic linking of static archives on linux
 SET (CMAKE_POSITION_INDEPENDENT_CODE ON)
