+++ src/commands/MathCommands.cc
@@ -112,7 +112,9 @@ solveCmdImpl(CommandHandler &data)
   const DoubleType charge_error = data.GetDoubleOption("charge_error");
   const DoubleType absolute_error = data.GetDoubleOption("absolute_error");
   const DoubleType relative_error = data.GetDoubleOption("relative_error");
+  const DoubleType maximum_error = data.GetDoubleOption("maximum_error");
   const int    maximum_iterations = data.GetIntegerOption("maximum_iterations");
+  const int    maximum_divergence = data.GetIntegerOption("maximum_divergence");
   const DoubleType frequency = data.GetDoubleOption("frequency");
   const std::string &outputNode = data.GetStringOption("output_node");
 
@@ -121,6 +123,8 @@ solveCmdImpl(CommandHandler &data)
   solver.SetRelError(relative_error);
   solver.SetQRelError(charge_error);
   solver.SetMaxIter(maximum_iterations);
+  solver.SetMaxDiv(maximum_divergence);
+  solver.SetMaxAbsError(maximum_error);
 
   std::unique_ptr<dsMath::LinearSolver<DoubleType>> linearSolver;
 
@@ -213,7 +217,9 @@ solveCmd(CommandHandler &data)
     {"type",               "", dsGetArgs::optionType::STRING, dsGetArgs::requiredType::REQUIRED, stringCannotBeEmpty},
     {"absolute_error",     "0", dsGetArgs::optionType::FLOAT, dsGetArgs::requiredType::OPTIONAL},
     {"relative_error",     "0", dsGetArgs::optionType::FLOAT, dsGetArgs::requiredType::OPTIONAL},
+    {"maximum_error",      "MAXDOUBLE", dsGetArgs::optionType::FLOAT, dsGetArgs::requiredType::OPTIONAL},
     {"maximum_iterations", "20", dsGetArgs::optionType::INTEGER, dsGetArgs::requiredType::OPTIONAL},
+    {"maximum_divergence", "20", dsGetArgs::optionType::INTEGER, dsGetArgs::requiredType::OPTIONAL},
     {"frequency",    "0.0", dsGetArgs::optionType::FLOAT, dsGetArgs::requiredType::OPTIONAL},
     {"output_node",  "", dsGetArgs::optionType::STRING, dsGetArgs::requiredType::OPTIONAL},
     {"solver_type",  "direct", dsGetArgs::optionType::STRING, dsGetArgs::requiredType::OPTIONAL},
