### 2021-11-21 d0f0e52
update release notes, version number, and pydoc
