+++ src/pythonapi/DevsimDoc.cc
@@ -1823,7 +1823,7 @@ R"(    devsim.vector_element_model (device, region, element_model)
     region : str
        The selected region
     element_model : str
-       The element model for which we are calculating the vector compoenents
+       The element model for which we are calculating the vector components
 
     Notes
     -----
@@ -1907,7 +1907,7 @@ R"(    devsim.get_matrix_and_rhs (format)
 )";
 
 static const char solve_doc[] =
-R"(    devsim.solve (type, solver_type, absolute_error, relative_error, charge_error, gamma, tdelta, maximum_iterations, frequency, output_node, info)
+R"(    devsim.solve (type, solver_type, absolute_error, relative_error, maximum_error, charge_error, gamma, tdelta, maximum_iterations, maximum_divergence, frequency, output_node, info)
 
     Call the solver.  A small-signal AC source is set with the circuit voltage source.
 
@@ -1921,6 +1921,8 @@ R"(    devsim.solve (type, solver_type, absolute_error, relative_error, charge_e
        Required update norm in the solve (default 0.0)
     relative_error : Float, optional
        Required relative update in the solve (default 0.0)
+    maximum_error : Float, optional
+       Maximum absolute error before solve stops (default MAXDOUBLE)
     charge_error : Float, optional
        Relative error between projected and solved charge during transient simulation (default 0.0)
     gamma : Float, optional
@@ -1929,6 +1931,8 @@ R"(    devsim.solve (type, solver_type, absolute_error, relative_error, charge_e
        time step (default 0.0)
     maximum_iterations : int, optional
        Maximum number of iterations in the DC solve (default 20)
+    maximum_divergence : int, optional
+       Maximum number of diverging iterations during solve (default 20)
     frequency : Float, optional
        Frequency for small-signal AC simulation (default 0.0)
     output_node : str, optional
