+++ CMakeLists.txt
@@ -10,7 +10,7 @@ set (CMAKE_CXX_STANDARD 17)
 INCLUDE(${PROJECT_SOURCE_DIR}/cmake/${DEVSIM_CONFIG}.cmake)
 INCLUDE(${PROJECT_SOURCE_DIR}/local.cmake OPTIONAL)
 
-ADD_DEFINITIONS(-DDEVSIM_VERSION_STRING="\\"1.6.0\\"")
+ADD_DEFINITIONS(-DDEVSIM_VERSION_STRING="\\"1.6.1\\"")
 ADD_DEFINITIONS(-DDEVSIM_COPYRIGHT_YEAR="\\"2009-2021\\"")
 
 IF (MKL_PARDISO)
