### 2021-11-24 devsim/devsim#76 @dalonsoa
[JOSS Review] Installation instructions (MacOS) are incomplete
**Part of the review for publication in [JOSS](https://github.com/openjournals/joss-reviews/issues/3898)**

[Installation instructions](https://github.com/devsim/devsim/blob/main/INSTALL) are incomplete. They just say - in esence - "use this script if you use MacOS" or "use this script if you use Centos", etc. But not how to actually used those scripts. In particular, if you just do:

```bash
$ sh scripts/build_macos.sh
```

It fails with:

```output
ERROR: FIRST ARGUMENT MUST BE gcc OR clang
```

because it is necessary to provide the compiler as an input argument to the script, but that is not mentioned anywhere. Input parameters for those scripts should be desrcribed.

When it runs, it creates a conda environment and downloads and install a bunch of dependencies - what happens if there is no `conda`? - and, in my case, it finally fails with:

```output
CMake Error: The source directory "/Users/dalonsoa/Documents/REVIEWS/devsim/devsim/external/symdiff" does not appear to contain CMakeLists.txt
```

`symdiff` seems to be another dependency embeded within the `devsim` directory tree that is not really downloaded when doing `git clone`, and therefore `CMakeLists.txt` is indeed missing. If the method for getting `symdiff` is another one, that should be described.

While having the binaries is great and will be more than enough for most users, most of the time, having a more complete installation instructions will be really helpful, specially for possible contributors and developers who will need to use the source.

Also, it will be nice if the INSTALL file were in Markdown or other format that renders nicely.
