### 2021-11-24 devsim/devsim#77 @dalonsoa
[JOSS Review] Instructions for using MacOS binaries need update
**Part of the review for publication in [JOSS](https://github.com/openjournals/joss-reviews/issues/3898)**

When following the instructions to setup `devsim` binaries on MacOS, I get two issues that should be described in the instructions:

When doing:

```python
import devsim
```

I get a message saying that the `.so` file is not authorised. I need to authorize the use of libraries from unknown developers (at least in MacOS Monterrey 12.01). This is similar to the issue - and solution - described [here](https://github.com/SAP/node-rfc/issues/140#issuecomment-675260978).

After doing that, then it complains because MKL is not found.

```output
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/dalonsoa/Documents/REVIEWS/devsim/macos/lib/devsim/__init__.py", line 8, in <module>
    from .devsim_py3 import *
ImportError: dlopen(/Users/dalonsoa/Documents/REVIEWS/devsim/macos/lib/devsim/devsim_py3.so, 0x0002): Library not loaded: @rpath/libmkl_rt.dylib
  Referenced from: /Users/dalonsoa/Documents/REVIEWS/devsim/macos/lib/devsim/devsim_py3.so
  Reason: tried: '/usr/lib/libmkl_rt.dylib' (no such file)
```

I believe [instructions for MacOS](https://github.com/devsim/devsim/blob/c6ee02f1ab553f65754752c370ae3dc1c1548224/macos.txt) are slightly misleading. Having MKL **is required**, not optional as it is sugetsed in the documentation, which states that "one of these packages" in reference to Python 3 and Intel MKL. Both are needed.

Once MKL is installed, all works as expected.

Also, it will be nice if the different binary installation files were in Markdown or other format that renders nicely.
