### 2021-11-24 devsim/devsim#78 @dalonsoa
[JOSS Review] Contributing Guidelines are outdated
Part of the review for publication in [JOSS](https://github.com/openjournals/joss-reviews/issues/3898)

The [contributing guidelines](https://github.com/devsim/devsim/blob/c6ee02f1ab553f65754752c370ae3dc1c1548224/CONTRIBUTING.md) need to be updated - and possibly expanded with more informatin. They point to a [mailing list](https://groups.google.com/g/devsim?pli=1), but that one was closed in April and now a [Forum](https://forum.devsim.org/) must be used.

They are also quite short and limitted, and not asnwering explicitly how to:
1. Contribute to the software
2. Report issues or problems with the software
3. Seek support

All they say is "go to the forum".

I would say that using a forum is not too standard for open source projects - using the issue tracker is more normal - but if it works withinn that communnity, I guess it is fine.
