### 2021-11-24 #3898 @dalonsoa
I have a problem with testing.

`devsim` includes a collection of python scripts that run as part of its testing suite and that check the different functionality. However, it uses no standard testing framework for the high level python interface ([`pytest`](https://docs.pytest.org/en/6.2.x/), [`unittest`](https://docs.python.org/3/library/unittest.html)) nor a unit testing framewrok for the low level c++ functinality ([`catch`](https://github.com/catchorg/Catch2)). What there is now is neither exaustive nor sustainable.

However, requiring to implement any of this is a major and very time consuming undertaking and I am reluctant to block the publication until it is done.

@lucydot , what si your suggestion on this?
