### 2021-11-24 devsim/devsim#79 @dalonsoa
[JOSS Review] Confusing multiple documentation sources
Part of the review for publication in [JOSS](https://github.com/openjournals/joss-reviews/issues/3898)

It is quite disconcerting that there are two separate documentation sources:
- [DEVSIM TCAD Semiconductor Device Simulator](https://devsim.org/) in https://devsim.org
- [DEVSIM Manual](https://devsim.net/) in https://devsim.net/

The second one is very complete, although although, personally, I would move the release notes towards the end. I am unsure of the purpose of the first one.

I think it will be important, for the sake of clarity, to colapse the above in just one documentation source and  that the entry page is a clear, high level description of what problems the software is designed to solve and who the target audience is. This should also be included in the README file.
