### 2021-11-24 #3898 @tcaduser
>

Hello,

The tests are driven via CTest and are run as part of the release system.  For each platform, the regressions are run and new results are updated before each release.

These are the existing regression tests:
https://github.com/devsim/devsim_tests_msys
https://github.com/devsim/devsim_tests_win64
https://github.com/devsim/devsim_tests_macos_gcc_x86_64
https://github.com/devsim/devsim_tests_linux_x86_64

This includes the tests in the ``testing`` and examples in the ``examples`` directory.
