### 2021-11-24 #3898 @dalonsoa
About the paper:

The description of the **State of the field** is a bit limited. It mentions two open source packages that operate in the same field, but it does not make any comparison with them in terms of features or capabilities. More importantly, it does not compare with the very strong comercial packages (SILVACO, COMSOL). Obviously, one of the benefits of `devsim` is that is is open source and free compared with the $$$$ that the comercial solutions cost, so I think there is a strong case to be made on how this open source tool can leverage the functionality of some of the other, way less accessible, commercial packages.
