### 2021-11-25 52e575d
update for JOSS publication
