+++ paper/paper.bib
@@ -17,6 +17,20 @@
   doi={10.1109/TED.2021.3094776},
 }
 
+@article{Lauwaert_2021,
+	doi = {10.1088/1361-6463/ac34ad},
+	url = {https://doi.org/10.1088/1361-6463/ac34ad},
+	year = 2021,
+	month = {nov},
+	publisher = {{IOP} Publishing},
+	volume = {55},
+	number = {8},
+	pages = {085101},
+	author = {Johan Lauwaert},
+	title = {Technology computer aided design based deep level transient spectra: simulation of high-purity germanium crystals},
+	journal = {Journal of Physics D: Applied Physics}
+}
+
 
 @article{chen:pssa,
 author = {Chen, Qiusong and Lin, Dong and Wang, Qinhao and Yang, Jiang and Sanchez, Juan E. and Zhu, Guodong},
