### 2021-11-25 devsim/devsim#76 @tcaduser
Hello @dalonsoa,

Thank you for your suggestions.  I have converted INSTALL to INSTALL.md in commit 52e575d606e.  In this new file I have updated the instructions for installing the binary, and compiling the application from source.  The symdiff project is indeed required, and is included into the project as a submodule.  I have made this requirement more explicit, and included instructions for automatically updating the submodule..
