### 2021-11-25 devsim/devsim#77 @tcaduser
Hello @dalonsoa,
Thank you for your comments.  I have updated the instructions for installing the binaries in INSTALL.md for commit 52e575d606e.  This now includes the requirement for Intel MKL.

My mac laptop is based on Mojave (Version 10.14.6).  I am not familiar with this restriction for modules loaded via Anaconda.  I have seen this issue for modules loaded via the system Python 2.7 interpreter in Mojave, since it is an executable attempting to load a shared library in a non-system directory.

Is your Anaconda installation located in your home directory, or in a system directory?  At the moment, I do not have access to Monterrey 12.01 system.
