### 2021-11-25 devsim/devsim#78 @tcaduser
Hi @dalonsoa,

Thank you for your comments.  With commit 52e575d606e, I have updated CONTRIBUTING.md to point to the new forum.  My understanding of a contributing file, is that it is for new development.  I was not aware that it is instructions concerning support.  Should I be moving the Support section of README.md into CONTRIBUTING.md?  Your help is greatly appreciated.
