### 2021-11-25 devsim/devsim#79 @tcaduser
The purpose of the first website (devsim.org) is to be the homepage for the project.  The second website is the documentation, and is generated directly from: https://github.com/devsim/devsim_documentation.  For this reason, I would like to keep them separate.

I have tried to state in the preamble of https://devsim.net the difference between these websites.

Since https://devsim.org is the homepage for the project, it seems this place might be the best to place the motivation for this project.
