+++ dist/package_linux.sh
@@ -44,7 +44,7 @@ mkdir -p ${DIST_DIR}/doc
 cp ../doc/devsim.pdf ${DIST_DIR}/doc
 cp ${SYMDIFF_DOCUMENTATION_DIR}/symdiff.pdf ${DIST_DIR}/doc
 
-for i in INSTALL.md NOTICE LICENSE linux.txt README.md CHANGES.md install.py; do
+for i in INSTALL.md NOTICE LICENSE README.md CHANGES.md install.py; do
 cp ../$i ${DIST_DIR}
 done
 
