+++ macos.txt
@@ -1,39 +1,0 @@
-This is the DEVSIM distribution.  Extract the files from
-the archive into your preferred directory.  Product
-documentation is available in the devsim/doc directory.
-
-Default Version:
-
-macOS provides all of the necessary dependencies to run DEVSIM from the built-in python interpreter.
-
-GCC Version:
-
-DEVSIM depends on having one of the following 64 bit applications installed on your system:
-
-* Python 3 (3.6 or higher)
-* Intel MKL
-
-All of these dependencies are available by installing:
-Anaconda Python
-https://www.anaconda.com/products/individual#Downloads
-
-or Miniconda Python
-https://conda.io/miniconda.html
-
-It is necessary to install these packages into your conda environment:
-conda install mkl numpy
-
-Setting paths:
-csh:
-  setenv PYTHONPATH /path/to/devsim/lib
-
-bash:
-  export PYTHONPATH=/path/to/devsim/lib
-
-Running:
-The module can then be run by importing the devsim module from your python script.
-
-Issues:
-If you have problems running the software, please report an issue at:
-https://forum.devsim.org
-
