+++ CHANGES.md
@@ -1,14 +1,16 @@
-### Version 1.6.1
+# CHANGES
 
-#### Versioned MKL DLL in release build
+## Version 1.6.1
+
+### Versioned MKL DLL in release build
 
 The Intel Math Kernel Library now uses versioned library names.  Binary releases are now updated to against the latest versioned dll names from MKL available in the Anaconda Python distribution.
 
-#### Fixed issue in ramp function.
+### Fixed issue in ramp function.
 
 The ``rampbias`` function in the ``devsim.python_packages.ramp`` module has been fixed to properly reduce the bias when there is a convergence failure.
 
-#### Transient Simulation
+### Transient Simulation
 
 Fixed bug with ``transient_tr`` (trapezoidal) time integration method where the wrong sign was used to integrate previous time steps.
 
@@ -16,49 +18,53 @@ Fixed bug in the charge error calculation, which calculates the simulation resul
 
 Added ``testing/transient_rc.py`` test which compares simulation with analytic result for RC circuit.
 
-#### Solver
+### Solver
 
-##### Convergence Tests
+#### Convergence Tests
 
 The ``maximum_error`` and ``maximum_divergence`` options where added to the ``solve`` command.  If the absolute error of any iteration goes above ``maximum_error``, the simulation stops with a convergence failure.  The ``maximum_divergence`` is the maximum number of iterations that the simulator error may increase before stopping.
 
-##### Verbosity
+#### Verbosity
 
 During the ``solve``, circuit node and circuit solution information is no longer printed to the screen for the default verbosity level.  In addition, the number of equations per device and region is no longer displayed at the start of the first iteration.
 
-##### SuperLU
+#### SuperLU
 
 The code now supports newer versions of ``SuperLU``.  The release version is still using SuperLU 4.3 for the iterative solution method, and the Intel MKL Pardiso for the direct solve method.
 
-#### Build Scripts
+### Build Scripts
 
 The build scripts have been updated on all platforms to be less dependent on specific Python 3 versions.
 
 An updated fedora build script has been added.  It uses the system installed ``SuperLU`` as the direct solver.
 
-### Version 1.6.0
+### Documentation Files
 
-#### Array Type Input and Output
+Some out of date files (e.g. RELEASE, INSTALL, . . .) have been removed.  The [README.md](README.md) has been updated and the [INSTALL.md](INSTALL.md) have been updated.
+
+## Version 1.6.0
+
+### Array Type Input and Output
 
 In most circumstances, the software now returns numerical data using the Python ``array`` class.  This is more efficient than using standard lists, as it encapsulates a contiguous block of memory.  More information about this class can be found at [https://docs.python.org/3/library/array.html](https://docs.python.org/3/library/array.html).  The representation can be easily converted to lists and ``numpy`` arrays for efficient manipulation.
 
 When accepting user input involving lists of homogenous data, such as ``set_node_values`` the user may enter data using either a list, string of bytes, or the ``array`` class.  It may also be used to input ``numpy`` arrays or any other class with a ``tobytes`` method.
 
-#### Get Matrix and RHS for External Use
+### Get Matrix and RHS for External Use
 
 The ``get_matrix_and_rhs`` command has been added to assemble the static and dynamic matrices, as well as their right hand sides, based on the current state of the device being simulated.  The ``format`` option is used to specify the sparse matrix format, which may be either in the compressed column or compressed row formats, ``csc`` or ``csr``.
 
-#### Maximum Divergence Count
+### Maximum Divergence Count
 
[…]
