### 2021-11-25 devsim/devsim#77 @dalonsoa
Yes, `conda` is in my home director (in `~/opt/miniconda3/bin/conda`, in particular).

Who knows about the restriction! Apple adds more and more restrictions with each release of MacOS. I am not surprised at all that it works in Mojave but not in Monterry. I would mention in the installation instructions that this issue *might* hapen, as well as a possible solution. It could be in a Troubleshoot section or something like that.
