### 2021-11-25 devsim/devsim#78 @dalonsoa
You are absolutely right: The contributing file is about how to contirbute to the software. For support, asking to open an issue - or go to the forum in your case - is enough.

Nevertheless, I do believe the CONTRIBUTING file needs improvement to help possible contirbutors learn what tools should be used, how the code should be structured, etc. For example, what are the expectations about testing new code? Or what should a new feature be placed? They cal always ask in the forum, but it helps if that information is offered upfront. You can have a look at the [CONTRIBUTING file for the PyBaMM project](https://github.com/pybamm-team/PyBaMM/blob/develop/CONTRIBUTING.md) (I feel that one is excessive, to be honest, but it is a good example).

Now, I see you are, pretty much, the sole contributor to the project. If you want to keep it that way for waheter reason, that is absolutely fine, but then you must make that clear: that you do not accept external contributors. There is noting wrong about that - `sqlite` for example follows that model; check [section about "Open-source, not Open-contribution" in their copyrigth page](https://sqlite.org/copyright.html). But it has to be clear upfront.
