### 2021-11-26 #3898 @tcaduser
@dalonsoa
> About the paper:
>
> The description of the **State of the field** is a bit limited. It mentions two open source packages that operate in the same field, but it does not make any comparison with them in terms of features or capabilities. More importantly, it does not compare with the very strong comercial packages (SILVACO, COMSOL). Obviously, one of the benefits of `devsim` is that is is open source and free compared with the $$$$ that the comercial solutions cost, so I think there is a strong case to be made on how this open source tool can leverage the functionality of some of the other, way less accessible, commercial packages.

Thank you for your insight.  The TCAD landscape is very broad.  Since I have not used the other OSI-compliant open source offerings, I do not feel it is fair to comment on their capabilities.  Especially in that both Solcore and Charon are subject to rapid change.  Cogenda TCAD (GPL3) has not made any updates to their software in the recent future.

As for the commercial offering, I would find it difficult to compare this tool directly to the offerings of Synopsys, Silvaco, and COMSOL.  From what I can tell, the first two are comprehensive solutions with person decades of experience.  COMSOL is an FEM multiphysics package, that made the extra effort to implement FVM to incorporate Scharfetter Gummel.

I would prefer to not even get started discussing the tools with restrictive closed-source academic licenses.  Stanford's PISCES, Open NESS come to mind.

How much detail would you want for this section?  I have added an additional citation to the paper for a researcher using DEVSIM for DLTS.
