### 2021-11-26 devsim/devsim#78 @tcaduser
@dalonsoa

I welcome contributions, but since there have been few, the process has not been formalized.

Most of the contributions to the software has been through bug reporting.  Some users of the software have shared their physics models or simulation structures in Python as independent projects, or documented them as journal papers.  I refer to them on the project webpage as I am made aware of them.
https://devsim.org/introduction.html#using-the-simulator

As far as changes to the core application, I have had 2 people contribute build scripts for their platform and 1 person provide a patch for a newer version of SuperLU.  One person provide a python prototype for the "Gauss Fermi" integral, which I translated to C++.
