### 2021-11-26 devsim/devsim#77 @tcaduser
I will add a note to the INSTALL.md
