### 2021-11-26 f151837
add note about macOS
