### 2021-11-28 e80279b
Use ADD_COMPILE_OPTIONS from cmake
