+++ cmake/ubuntu_18.04.cmake
@@ -1,5 +1,5 @@
-ADD_DEFINITIONS(-fvisibility=hidden)
-ADD_COMPILE_OPTIONS(-Wall)
+ADD_COMPILE_OPTIONS(-Wall -fvisibility=hidden)
+
 SET (FLEX /usr/bin/flex)
 SET (BISON /usr/bin/bison)
 SET (EXTERNAL_LIB ${PROJECT_SOURCE_DIR}/external)
