### 2021-11-28 680abcd
write interface as nodes if not provided as elements
