### 2021-11-28 dd45fee
create_interface_from_nodes
