+++ src/commands/MeshingCommands.cc
@@ -1153,6 +1153,125 @@ createContactFromInterfaceCmd(CommandHandler &data)
     data.SetEmptyResult();
 }
 
+void createInterfaceFromNodesCmd(CommandHandler &data)
+{
+  std::string errorString;
+
+  const std::string commandName = data.GetCommandName();
+
+  using namespace dsGetArgs;
+  static dsGetArgs::Option option[] =
+  {
+      {"device",  "", dsGetArgs::optionType::STRING, dsGetArgs::requiredType::REQUIRED, mustBeValidDevice},
+      {"name",    "", dsGetArgs::optionType::STRING, dsGetArgs::requiredType::REQUIRED, stringCannotBeEmpty},
+      {"region0", "", dsGetArgs::optionType::STRING, dsGetArgs::requiredType::REQUIRED, mustBeValidRegion},
+      {"region1", "", dsGetArgs::optionType::STRING, dsGetArgs::requiredType::REQUIRED, mustBeValidRegion},
+      {"nodes0",  "", dsGetArgs::optionType::LIST,   dsGetArgs::requiredType::REQUIRED, nullptr},
+      {"nodes1",  "", dsGetArgs::optionType::LIST,   dsGetArgs::requiredType::REQUIRED, nullptr},
+      {nullptr,  nullptr, dsGetArgs::optionType::STRING, dsGetArgs::requiredType::OPTIONAL}
+  };
+
+  dsGetArgs::switchList switches = nullptr;
+
+
+  bool error = data.processOptions(option, switches, errorString);
+
+  if (error)
+  {
+      data.SetErrorResult(errorString);
+      return;
+  }
+
+  const std::string &deviceName = data.GetStringOption("device");
+  const std::string &name       = data.GetStringOption("name");
+
+  std::string regionNames[2];
+  regionNames[0] = data.GetStringOption("region0");
+  regionNames[1] = data.GetStringOption("region1");
+
+  Device *dev = nullptr;
+  Region *reg[2] = {};
+
+  for (size_t i = 0; i < 2; ++i)
+  {
+    errorString += ValidateDeviceAndRegion(deviceName, regionNames[i], dev, reg[i]);
+  }
+
+  if (!errorString.empty())
+  {
+      data.SetErrorResult(errorString);
+      return;
+  }
+
+
+  std::vector<size_t> nodes[2];
+  for (size_t i = 0; i < 2; ++i)
+  {
+    std::string oname("nodes");
+    oname += std::to_string(i);
+    ObjectHolder vdata = data.GetObjectHolder(oname);
+    if (vdata.IsList())
+    {
+      bool ok = vdata.GetUnsignedLongList(nodes[i]);
+      if (!ok)
+      {
+        std::ostringstream os;
+        os << "Option \"" << oname << "\" could not be converted to a list of unsigned integers\n";
+        errorString += os.str();
+      }
+    }
+  }
+
+  if (!errorString.empty())
+  {
+      data.SetErrorResult(errorString);
+      return;
+  }
+
+  if ((nodes[0].empty() || nodes[1].empty()) || (nodes[0].size() != nodes[1].size()))
+  {
+    std::ostringstream os;
+    os << "Node lists \"nodes0\", \"nodes1\" are empty or not the same size " << nodes[0].size() << " " << nodes[1].size() <<"\n";
+    errorString = os.str();
+    data.SetErrorResult(errorString);
+    return;
+  }
+
+  size_t error_count = 0;
+  std::vector<ConstNodePtr> inodes[2];
+  for (size_t i = 0; i < 2; ++i)
+  {
+    inodes[i].resize(nodes[i].size(), nullptr);
[…]
