+++ src/commands/ModelCommands.cc
@@ -2290,8 +2290,3 @@ Commands ModelCommands[] = {
 
 }
 
-//TODO:  "get dependency list of models"
-//TODO:  "Evaluate expression"
-
-//TODO: set_interface_value
-
