### 2021-11-29 c19234d
set_initial_condition
