+++ src/pythonapi/DevsimDoc.cc
@@ -888,10 +888,6 @@ R"(    devsim.create_contact_from_interface (device, region, interface, material
        name of the contact begin created
 )";
 
-static const char create_interface_from_nodes_doc[] =
-R"(
-)";
-
 static const char create_device_doc[] =
 R"(    devsim.create_device (mesh, device)
 
@@ -949,6 +945,27 @@ R"(    devsim.create_gmsh_mesh (mesh, file, coordinates, elements, physical_name
 
       - Each node of the element indexes into the coordinates list.
 
+)";
+
+static const char create_interface_from_nodes_doc[] =
+R"(    devsim.create_interface_from_nodes (device, name, region0, region1, nodes0, nodes1)
+
+    Creates an interface from lists of nodes
+
+    Parameters
+    ----------
+    device : str
+       The selected device
+    name : str
+       name of the interface begin created
+    region0 : str
+       first region that the interface is attached to
+    region1 : str
+       second region that the interface is attached to
+    nodes0 : str
+       list of nodes for the interface in the first region
+    nodes1 : str
+       list of nodes for the interface in the second region
 )";
 
 static const char finalize_mesh_doc[] =
@@ -1911,9 +1928,17 @@ R"(    devsim.get_matrix_and_rhs (format)
 )";
 
 static const char set_initial_condition_doc[] =
-R"(
-)";
+R"(    devsim.set_initial_condition (static_rhs, dynamic_rhs)
 
+    Sets the initial condition for subsequent transient solver steps.
+
+    Parameters
+    ----------
+    static_rhs : list, optional
+       List of double values for non time-displacement terms in right hand side.
+    dynamic_rhs : list, optional
+       List of double values for time-displacement terms in right hand side.
+)";
 
 static const char solve_doc[] =
 R"(    devsim.solve (type, solver_type, absolute_error, relative_error, maximum_error, charge_error, gamma, tdelta, maximum_iterations, maximum_divergence, frequency, output_node, info)
