### 2021-11-29 fc32643
Update build scripts and instructions remove unused switch option update doc
