+++ scripts/build_centos_8.sh
@@ -15,14 +15,8 @@ export F77="/usr/bin/gfortran"
 
 #minimal conda environments to prevent linking against the wrong libraries
 cd ${HOME}
-if [ ! -f Miniconda3-latest-Linux-x86_64.sh ]
-then
-curl -L -O https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh;
-bash Miniconda3-latest-Linux-x86_64.sh -b -p ${HOME}/anaconda;
-${HOME}/anaconda/bin/conda create  -y --name python3_devsim_build python=3
-${HOME}/anaconda/bin/conda install -y --name python3_devsim_build mkl mkl-devel mkl-include boost cmake
-fi
-source ${HOME}/anaconda/bin/activate python3_devsim_build
+conda create  -y --name python3_devsim_build python=3 mkl mkl-devel mkl-include boost cmake
+source activate python3_devsim_build
 
 export PYTHON3_BIN=python
 export PYTHON3_INCLUDE=$(python -c "from sysconfig import get_paths as gp; print(gp()['include'])")

[…]
