### 2021-11-30 #3898 @lucydot
RE: testing - @dalonsoa are the queries you raised partly answered by the response from @tcaduser (which seems to relate to a C++ testing unit framework) or are you still hoping for more clarification from editorial point of view?

Ideally the Python interface would have automated testing using pytest or similar, however if there are tests that a user can run manually then this is acceptable.

@tcaduser It may be that you need to clarify details around testing in the documentation (indicate coverage, what is automated, what is manual, how a user can verify that the installation is correct and the software is working as expected).
