### 2021-12-01 0269af7
update for variable_name option removal
