+++ examples/capacitance/cap2d.py
@@ -118,7 +118,6 @@ for c in ("top", "bot"):
                        equation="1")
 
     contact_equation(device=device, contact=c, name="PotentialEquation",
-                     variable_name="Potential",
                      node_model="%s_bc" % c, edge_charge_model="DField")
 
 ###
