### 2021-12-01 #3898 @jgoldfar
> 👋 @jgoldfar, please update us on how your review is going (this is an automated reminder).

I am going through the steps in [this listing](https://github.com/openjournals/joss-reviews/issues/3898#issue-1047430015) in this issue, and have found that I cannot toggle the "check boxes" - is this a problem on my end, or is it because I am not assigned to this issue?
