### 2021-12-01 #3898 @tcaduser
@jgoldfar, @dalonsoa

I have merged the ``joss`` branch  with the ``r1.7.0`` branch for the latest versions of INSTALL.md, README.md, . . .  The pdf form of the documentation is in ``doc/devsim.pdf`` of the same branch.
