### 2021-12-01 0a597c2
add link to documentation source to README.md
