### 2021-12-11 #3898 @tcaduser
@jgoldfar, @dalonsoa

The newest release branch ``r2.0.0`` has been merged into the ``joss`` branch.
