+++ CMakeLists.txt
@@ -1,4 +1,6 @@
 PROJECT (DEVSIM CXX)
+cmake_minimum_required(VERSION 3.0)
+
 OPTION(VTKWRITER    "Build with VTK Writer" ON)
 OPTION(PYTHON3      "Build Python 3 Interpreter" ON)
 OPTION(DEVSIM_EXTENDED_PRECISION "Build with extended precision" OFF)
@@ -10,8 +12,13 @@ set (CMAKE_CXX_STANDARD 17)
 INCLUDE(${PROJECT_SOURCE_DIR}/cmake/${DEVSIM_CONFIG}.cmake)
 INCLUDE(${PROJECT_SOURCE_DIR}/local.cmake OPTIONAL)
 
+<<<<<<< HEAD
 ADD_DEFINITIONS(-DDEVSIM_VERSION_STRING="\\"2.0.0\\"")
 ADD_DEFINITIONS(-DDEVSIM_COPYRIGHT_YEAR="\\"2009-2021\\"")
+=======
+ADD_DEFINITIONS(-DDEVSIM_VERSION_STRING=\"1.7.0\")
+ADD_DEFINITIONS(-DDEVSIM_COPYRIGHT_YEAR=\"2009-2021\")
+>>>>>>> 7c59273... reduce warnings
 
 IF (MKL_PARDISO)
 ADD_DEFINITIONS(-DUSE_MKL_PARDISO)
