### 2021-12-11 8bc9fe6
update version string for cmake
