### 2021-12-16 #3898 @dalonsoa
@tcaduser , I'm happy with the improvements in the installation instructions - great job! - and I've also checked some features and performance. All good on that front. After the clarification above, I am also happy with the testing part - although I strongly recomend opening an issue about implementing a unit testing suite - for both the C++ part and the Python part. But that is work for the future.

Now, I am still not convinced by the Statements of Purpose in both the paper and the documentation, as well as the State of the Field. I understand the arguments you give [above](https://github.com/openjournals/joss-reviews/issues/3898#issuecomment-980085105), but even that discussion - without actually comparing features, something I agree is very challenging - is extremelly valuable to be written down as part of the documents. This way, people can easily put DEVISIM on the map of TCAD software, which is an enormous field.

Also, the Contributing guidelines. At the very least, I would describe the whole ecosystem of repositories you have in place hosting the source code, the tests for the different platforms, etc, so potential contributors cand find out how things are arranged.

In summary: the tool itself is great, but it is necessary some extra work to clarify where it sits in the big scheme of things and how to support it.
