+++ CONTRIBUTING.md
@@ -19,4 +19,4 @@ If you would like to participate in this project, please join our email forum:
 https://forum.devsim.org
 and let us know how you would like to help.
 
-
+Please see the [Testing](README.md#Testing) and [Related Projects](README.md#Related-Projects) section in [README.md](README.md) for additional places where help is needed.
