+++ paper/paper.bib
@@ -1,4 +1,28 @@
 
+@article{OSI:2007,
+  author = {OSI},
+  title = {Open Source Definition},
+  url = {http://www.opensource.org/docs/definition.php},
+  urldate = {December 18, 2021},
+  year = 2007
+}
+
+@book{Selberherr,
+author = {Siegfried Selberherr},
+title = {Analysis and simulation of semiconductor devices},
+year = {1984},
+   Publisher = {Springer-Verlag},
+   Address = {NY},
+}
+
+@misc{symdiff,
+  author = {{DEVSIM LLC}},
+  title = {symdiff},
+  date = 2021,
+  urldate = {2021},
+  url = {https://symdiff.org}
+}
+
 @Misc{devsim:home,
         author = {{DEVSIM LLC}},
 	title = {{DEVSIM} {TCAD} Semiconductor Device Simulator},
