+++ paper/paper.bib
@@ -1,10 +1,10 @@
 
 @article{OSI:2007,
-  author = {OSI},
+  author = {{Open Source Initiative}},
   title = {Open Source Definition},
   url = {http://www.opensource.org/docs/definition.php},
   urldate = {December 18, 2021},
-  year = 2007
+  year = 2021
 }
 
 @book{Selberherr,
