### 2021-12-18 #3898 @tcaduser
@dalonsoa please check the updated paper, to see if it meets your expectations.  It adds an extended discussion comparing the simulator to the commercial offerings.  It also sets a list of requirements for the class of TCAD problems it solves.  The paper now explicitly references SYMDIFF, the symbolic differentiation engine that is used by the simulator, and that is also maintained by our project.

On the joss branch
The CONTRIBUTING.md now points back to the README.md, where I have explicitly added the related projects.
