### 2021-12-21 #3898 @dalonsoa
@tcaduser , now this looks really, relaly good, both the paper and the new README. Many thanks for taking the time to implementing the changes.

All good from my side. After merging the JOSS branch, **I recomend accepting the work for its publication in JOSS**.
