### 2022-01-10 #3898 @jgoldfar
Thanks @lucydot - I'm able to check off the tasks for this review now.

@tcaduser - well done on this comprehensive work. I would support a more automated approach to testing as suggested by @dalonsoa for the continued robustness of your project; it looks like you have integration tests as mentioned [in your comment](https://github.com/openjournals/joss-reviews/issues/3898#issuecomment-978052530), though they don't appear to be automated. Unit testing for every component of your application is probably too much to ask, and can create unnecessary fragility - you are probably the best person to decide on right combination of unit tests and automated functional or integration tests.

Recognizing that the technical aspects of the work are more of a journey than a destination, I too recommend **accepting this work for publication in JOSS**.
