### 2022-01-11 #3898 @lucydot
Hi @jgoldfar that's great - thank you for your comments. If you are happy that the software passes criteria, can I ask you to tick off the appropriate boxes at the top of the issue? There are two unticked. Note that the "automated testing" can be ticked if there are manual steps outlined so that a user can verify that the paper is working as intended - which I believe, from the discussions I've seen, is the case here.
