### 2022-01-13 #3898 @lucydot
On behalf of the JOSS editorial team, thank you very much @jgoldfar and @dalonsoa for your reviews ✨

@tcaduser it looks like we are onto the final lap!

I have a few editorial suggestions for the JOSS article - to improve readability:

- `DEVSIM is a technology computer-aided design (TCAD) software` : --> re-word this to `DEVSIM is technology computer-aided design (TCAD) software` or perhaps `DEVSIM is a technology computer-aided design (TCAD) code` or `DEVSIM is a technology computer-aided design (TCAD) simulator`.
- `This capability is useful for calibration to model existing fabricated devices`: --> I'm not sure I fully understand the meaning here. Perhaps `It can be used to model existing, fabricated devices for calibration purposes`.
- `saving the number of` : --> `reducing the number of`.
- `Software with TCAD simulation`: --> perhaps `Software tools with TCAD simulation` (the former sounds not quite right to me as it is unclear whether it is a single piece of software of multiple pieces of software - my suggestion makes it clear that it is the plural).
- `The number of open source
offerings, meeting the open source definition (OSD) (Open Source Initiative, 2021), has been
growing (Alonso-Álvarez et al., 2018; Cogenda, 2016; Sandia National Laboratories, 2020)` : --> `The number of open source
offerings, as defined by the open source initiative (Open Source Initiative, 2021), is growing (Alonso-Álvarez et al., 2018; Cogenda, 2016; Sandia National Laboratories, 2020).`
- `To our knowledge, the only OSD simulators, having most of these features, are Solcore,` : --> `To our knowledge the only OSD simulators having most of these features are Solcore,`
- Paragraph beginning`A unique feature of the software is its scripting model interface. It symbolically evaluates expressions and their derivatives.... ` : --> `A unique feature of DEVSIM is its scripting model interface which symbolically evaluates expressions and their derivatives. In this sense, DEVSIM is tailored toward modeling advanced TCAD simulation models. The core engine is maintained as a separate project, SYMDIFF (DEVSIM LLC, 2021b).`

Once you make the necessary changes please let me know here - thanks!
