+++ CMakeLists.txt
@@ -12,8 +12,8 @@ set (CMAKE_CXX_STANDARD 17)
 INCLUDE(${PROJECT_SOURCE_DIR}/cmake/${DEVSIM_CONFIG}.cmake)
 INCLUDE(${PROJECT_SOURCE_DIR}/local.cmake OPTIONAL)
 
-ADD_DEFINITIONS(-DDEVSIM_VERSION_STRING=\"2.0.0\")
-ADD_DEFINITIONS(-DDEVSIM_COPYRIGHT_YEAR=\"2009-2021\")
+ADD_DEFINITIONS(-DDEVSIM_VERSION_STRING=\"2.0.1\")
+ADD_DEFINITIONS(-DDEVSIM_COPYRIGHT_YEAR=\"2009-2022\")
 
 IF (MKL_PARDISO)
 ADD_DEFINITIONS(-DUSE_MKL_PARDISO)
