+++ paper/paper.md
@@ -19,11 +19,11 @@ bibliography: paper.bib
 
 # Summary
 
-`DEVSIM` is a technology computer-aided design (TCAD) software for semiconductor device simulation.  By solving the equations for electric fields and current flow, it simulates the electrical behavior of semiconductor devices, such as transistors.  This capability is useful for calibration to model existing fabricated devices.  It is also possible to explore novel device structures and exotic materials, saving the number of costly and time consuming manufacturing iterations.
+`DEVSIM` is technology computer-aided design (TCAD) software for semiconductor device simulation.  By solving the equations for electric fields and current flow, it simulates the electrical behavior of semiconductor devices, such as transistors.  It can be used to model existing, fabricated devices for calibration purposes.  It is also possible to explore novel device structures and exotic materials, reducing the number of costly and time consuming manufacturing iterations.
 
 # Statement of need
 
-`DEVSIM` is TCAD device simulation package written in C++, with a Python front end.  It is capable of simulating 1-D, 2-D and 3-D structures with models describing advanced physical effects [@devsim:home].  Software with TCAD simulation capabilities have existed in commercial and academic research tools for quite some time.  The number of open source offerings, meeting the open source definition (OSD) [@OSI:2007], has been growing [@cogenda; @charon; @solcore].
+`DEVSIM` is TCAD device simulation package written in C++, with a Python front end.  It is capable of simulating 1-D, 2-D and 3-D structures with models describing advanced physical effects [@devsim:home].  Software tools with TCAD simulation capabilities have existed in commercial and academic research tools for quite some time.  The number of open source offerings meeting the open source definition (OSD), as defined by the Open Source Initiative [@OSI:2007], has been growing [@cogenda; @charon; @solcore].
 
 Due to the expanse of TCAD simulation algorithms and models, it is important to note that this software fits into the class of continuum PDE based solvers for drift-diffusion semiconductor simulation [@Selberherr].  `DEVSIM` is intended to compare directly with the commercial TCAD offerings, such as Sentaurus Device from Synopsys, or the Victory Device Simulator from Silvaco.  The identifying factors for this class of software are:
 
@@ -32,9 +32,9 @@ Due to the expanse of TCAD simulation algorithms and models, it is important to 
 * Solution of 1D, 2D, and 3D unstructured meshes
 * Advanced models for mobility and semiclassical approaches for quantum effects
 
-While `DEVSIM` is not as complete as the commercial offerings, the project strives to fulfill the gaps by developing an open source community. To our knowledge, the only OSD simulators, having most of these features, are Solcore, Genius Semiconductor Device Simulator, and Charon [@solcore;@cogenda;@charon].  There are other simulators which provide their source code under restrictive academic research licenses.
+While `DEVSIM` is not as complete as the commercial offerings, the project strives to fulfill the gaps by developing an open source community. To our knowledge the only OSD-licensed simulators having most of these features are Solcore, Genius Semiconductor Device Simulator, and Charon [@solcore;@cogenda;@charon].  There are other simulators which provide their source code under restrictive academic research licenses.
 
-A unique feature of the software is its scripting model interface.  It symbolically evaluates expressions and their derivatives.  The core engine is maintained as a separate project, `SYMDIFF` [@symdiff]. It is then tailored toward modeling advanced TCAD simulation models. 
+A unique feature of DEVSIM is its scripting model interface, which symbolically evaluates expressions and their derivatives. In this sense, DEVSIM is tailored toward scripting advanced TCAD simulation models.  The symbolic engine is maintained as a separate project, `SYMDIFF` [@symdiff].
 
 These expressions are evaluated on the mesh structure by the C++ engine.
[…]
