+++ paper/paper.bib
@@ -32,8 +32,8 @@ year = {1984},
 
 @ARTICLE{sanchez:ieee,
   author={Sanchez, Juan E. and Chen, Qiusong},
-  journal={IEEE Transactions on Electron Devices}, 
-  title={Element Edge Based Discretization for TCAD Device Simulation}, 
+  journal={IEEE Transactions on Electron Devices},
+  title={Element Edge Based Discretization for TCAD Device Simulation},
   year={2021},
   volume={68},
   number={11},
