### 2022-01-14 #3898 @tcaduser
@lucydot , thank you for your excellent suggestions.  I have incorporated your changes, with some minor additional edits.

I have also merged in the r2.0.1 branch to the joss branch.  This incorporates additional changes to README.md, CONTRIBUTING.md, INSTALL.md.
