### 2022-02-05 #3898 @tcaduser
Sorry for the delay.  I have a DOI number reserved in Zenodo, and I will attempt to make a tagged release in the next 7 days.

Is the doi number of the proof, ``10.21105/joss.03898``, the one that is going to be on the final paper?
