### 2022-02-08 0b20e52
note about updating Intel MKL
