### 2022-02-08 #3898 @tcaduser
> Excellent @tcaduser we are at the final steps now.
>
> Please can you do the following:
>
>* Make a tagged release of your software

Version 2.0.1
https://github.com/devsim/devsim/releases/tag/v2.0.1

>* Archive this release on Zenodo (or similar)

[10.5281/zenodo.5908666](https://zenodo.org/record/5908666)

>   * The archive should have exactly the same title and author list as your JOSS paper
> * Report the archive DOI and version number in this thread

Thank you,

Juan
