### 2022-02-10 #3898 @lucydot
Hello @openjournals/joss-eics I think this paper is ready for the final checks from the EiC team before publication...
