### 2022-02-10 #3898 @kthyng
@tcaduser a few items before we wrap up:
1. version and archive look good ✅
2. Can you check out those missing dois, unless you have already checked previously? They aren't always correct.
3. Check the capitalization in your references. For example, Alonso-Alvarez et al has "python" uncapitalized. You can preserve capitalization with {} around characters in your .bib file.
