### 2022-02-10 2aaf7ca
update DOI and capitalization in references
