+++ paper/paper.bib
@@ -13,11 +13,12 @@ title = {Analysis and simulation of semiconductor devices},
 year = {1984},
    Publisher = {Springer-Verlag},
    Address = {NY},
+doi={10.1007/978-3-7091-8752-4}
 }
 
 @misc{symdiff,
   author = {{DEVSIM LLC}},
-  title = {symdiff},
+  title = {{SYMDIFF}},
   date = 2021,
   urldate = {2021},
   url = {https://symdiff.org}
@@ -63,7 +64,6 @@ journal = {physica status solidi (a)},
 pages = {2100408},
 keywords = {depletion layer electric field, depolarization field, ferroelectric thin-film transistors, leakage current, simulation},
 doi = {10.1002/pssa.202100408},
-url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/pssa.202100408},
 year = {2021},
 }
 
@@ -72,7 +72,7 @@ year = {2021},
   title = "Designing a Simulator for an Electrically-Pumped Organic Laser Diode",
   year = "2019",
   school = "California Polytechnic State University, San Luis Obispo, CA",
-  url = {https://doi.org/10.15368/theses.2019.60},
+  doi = {10.15368/theses.2019.60},
 }
 
 @misc{hulbert:software,
@@ -103,7 +103,7 @@ year = {2021},
 
 @article{solcore,
   author = {D. Alonso-Álvarez and T. Wilson and P. Pearce and M. Führer and D. Farrell and N. Ekins-Daukes},
-  title = {Solcore: a multi-scale, Python-based library for modelling solar cells and semiconductor materials},
+  title = {Solcore: a multi-scale, {P}ython-based library for modelling solar cells and semiconductor materials},
   year = {2018},
   journal = "Journal of Computational Electronics",
   pages = {1099--1123},
