### 2022-02-10 #3898 @tcaduser
Hello @kthyng ,

In addition, The URL for Lucy Whalley in the left panel does not work correctly in macOS Preview.  It probably needs the `https://` prefix
