### 2022-02-11 #3898 @kthyng
@openjournals/dev Can you check the links in the produced pdfs? As the author said above, the links for Lucy and for the archive in the left panel of the pdf do not work.
