### 2022-02-15 #3898 @arfon
> https://github.com/orgs/openjournals/teams/dev Can you check the links in the produced pdfs? As the author said above, the links for Lucy and for the archive in the left panel of the pdf do not work.

Apologies for the confusion, a couple of issues were going on here:

1. The archive URL only works in the final proofs (they work here https://github.com/openjournals/joss-papers/pull/2966)
2. The URL associated with @lucydot's JOSS profile was missing the `http://` prefix. I've now fixed this.

Both URLs are now working (in the final proofs here:  https://github.com/openjournals/joss-papers/pull/2966)
