### 2022-02-15 #3898 @tcaduser
Thanks to everyone for their help.  Can someone give me guidance as to where to request the tagged languages for the software on the publication page?  Here, it is showing as Ruby, Python, Batchfile:
https://joss.theoj.org/papers/10.21105/joss.03898

It should be just Python and C++.
