### 2022-02-16 #3898 @arfon
> It should be just Python and C++.

I've fixed this for you manually. We simply take what GitHub's Linguist library tells us here.
